Move prompt into its own file.
